check code is correct

+++ polygon-yolov5/utils/loss.py
@@ -318,7 +318,8 @@ class Polygon_ComputeLoss:
         na, nt = self.na, targets.shape[0]  # number of anchors, targets
         tcls, tbox, indices, anch = [], [], [], []
         # gain has 11 elements: img id, class id, xyxyxyxy, anchor id
-        gain = torch.ones(11, device=targets.device)  # normalized to gridspace gain
+        # gain = torch.ones(11, device=targets.device)  # normalized to gridspace gain
+        gain = torch.ones(11, device=targets.device).long()
         ai = torch.arange(na, device=targets.device).float().view(na, 1).repeat(1, nt)  # same as .repeat_interleave(nt)
         targets = torch.cat((targets.repeat(na, 1, 1), ai[:, :, None]), 2)  # append anchor indices
